LayoutTests/editing/pasteboard/data-transfer-set-data-sanitize-url-when-copying-in-null-origin-expected.txt

This tests calling setData with a malformed URL in a null origin document. The malformed value should not be readable in another document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS urlReadInSameDocument is "http://webkit.org/b/🤔?x=8 + 6<"
PASS htmlReadInSameDocument is "testing"
PASS typesInSameDocument.includes("text/uri-list") is true
PASS typesInSameDocument.includes("text/html") is true
PASS JSON.stringify(itemsInSameDocument[0]) is "{\"kind\":\"string\",\"type\":\"text/uri-list\"}"
PASS JSON.stringify(itemsInSameDocument[1]) is "{\"kind\":\"string\",\"type\":\"text/html\"}"
PASS event.clipboardData.getData("url") is "http://webkit.org/b/%F0%9F%A4%94?x=8%20+%206%3C"
PASS event.clipboardData.types.includes("text/uri-list") is true
PASS event.clipboardData.types.includes("text/html") is true
PASS JSON.stringify(pastedItems[0]) is "{\"kind\":\"string\",\"type\":\"text/uri-list\"}"
PASS JSON.stringify(pastedItems[1]) is "{\"kind\":\"string\",\"type\":\"text/html\"}"
PASS successfullyParsed is true

TEST COMPLETE
